Fix testFindCompositeRelationWithJoin by using the ExpressionInterface
